+++ src/components/theme/application/applicationProvider.component.tsx
@@ -16,12 +16,18 @@ import { StyleProvider } from '../style/styleProvider.component';
 import { ThemeProviderProps } from '../theme/themeProvider.component';
 import { ModalPanel } from '../modal/modalPanel.component';
 
-interface ComponentProps {
+interface EvaRuntimeProcessingProps {
   mapping: SchemaType;
   customMapping?: CustomSchemaType;
 }
 
-export type ApplicationProviderProps = ComponentProps & ThemeProviderProps;
+interface EvaBuildtimeProcessingProps {
+  styles: ThemeStyleType;
+}
+
+type EvaProcessingProps = EvaRuntimeProcessingProps | EvaBuildtimeProcessingProps;
+
+export type ApplicationProviderProps = EvaProcessingProps & ThemeProviderProps;
 export type ApplicationProviderElement = React.ReactElement<ApplicationProviderProps>;
 
 interface State {
@@ -38,13 +44,18 @@ interface State {
  * @extends React.Component
  *
  * @property {SchemaType} mapping - Determines the mapping for basic components.
- * This is designed to be provided by developers team and can be imported from npm package (e.g. `@eva-design/eva`).
+ * This is designed to be provided from any `@eva-design/*` package (e.g. `@eva-design/eva`)
+ * If provided, will be merged with customMapping to bootstrap eva during the runtime.
  *
- * @property {CustomSchemaType} customMapping - Determines the customization mapping.
+ * @property {CustomSchemaType} customMapping - Determines the customized mapping.
  * This is merged with `mapping` property and designed to be used components customization.
  *
  * @property {ThemeType} theme - Determines the theme for basic components.
  * This is designed to be provided by developers team and can be imported from npm package (e.g. `@eva-design/eva`).
+ *
+ * @property {ThemeStyleType} styles - Determines the styles compiled by bootstrapping eva.
+ * If provided, will replace runtime styles calculation.
+ * Should be used with `@ui-kitten/metro-config` package.
  *
  * @property {ReactNode} children - Determines application root component.
  *
@@ -53,40 +64,36 @@ interface State {
  * ```
  * import React from 'react';
  * import { ApplicationProvider, Layout, Text } from '@ui-kitten/components';
- * import { mapping, light as lightTheme } from '@eva-design/eva';
+ * import * as eva from '@eva-design/eva';
  *
- * export default class App extends React.Component {
- *
- *   render() {
- *     return (
- *       <ApplicationProvider
- *         mapping={mapping}
- *         theme={lightTheme}>
- *         <Layout style={{ flex: 1 }}>
- *           <Text>Welcome to UI Kitten</Text>
- *         </Layout>
- *       </ApplicationProvider>
- *     );
- *   }
- * }
+ * export default () => (
+ *   <ApplicationProvider {...eva} theme={eva.light}>
+ *     <Layout style={{ flex: 1 }}>
+ *       <Text>Welcome to UI Kitten</Text>
+ *     </Layout>
+ *   </ApplicationProvider>
+ * );
  * ```
  */
 export class ApplicationProvider extends React.Component<ApplicationProviderProps, State> {
+
+  public state: State = {
+    styles: (this.props as EvaBuildtimeProcessingProps).styles,
+  };
 
   private schemaProcessor: SchemaProcessor = new SchemaProcessor();
 
   constructor(props: ApplicationProviderProps) {
     super(props);
-    const { mapping, customMapping } = this.props;
 
-    const styles: ThemeStyleType = this.createStyles(mapping, customMapping);
-
-    this.state = { styles };
+    if (!this.state.styles) {
+      const { mapping, customMapping } = this.props as EvaRuntimeProcessingProps;
+      this.state.styles = this.createStyles(mapping, customMapping);
+    }
   }
 
   private createStyles = (mapping: SchemaType, custom: CustomSchemaType): ThemeStyleType => {
     const customizedMapping: SchemaType = merge({}, mapping, custom);
-
     return this.schemaProcessor.process(customizedMapping);
   };